requirements.txt
gf4
python-dotenv
vosk
yt_dlp
pydub
aiogram
requests